backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
psycopg==3.1.20  # Uncomment this line
psycopg-binary==3.2.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
# Add this line to your existing requirements.txt
Pillow==10.2.0
